Event: improve API to handle different event types
This patch makes the methods of the event object works in a
different way. The new API to register an event is as follows:

int duda_event_add(int sockfd, struct duda_request *dr,
                   int init_mode, int behavior,
                   int (*cb_on_read) (int, struct duda_request *),
                   int (*cb_on_write) (int, struct duda_request *),
                   int (*cb_on_error) (int, struct duda_request *),
                   int (*cb_on_close) (int, struct duda_request *),
                   int (*cb_on_timeout) (int, struct duda_request *));

So it introduce the new parameters init_mode and behavior, where
init_mode takes one of the following values:

    - DUDA_EVENT_READ
    - DUDA_EVENT_WRITE
    - DUDA_EVENT_RW
    - DUDA_EVENT_SLEEP

and behavior can be one of the epoll based modes:

    - DUDA_EVENT_LEVEL_TRIGGERED
    - DUDA_EVENT_EDGE_TRIGGERED

Also a new method to change the mode and behavior of a registered file
descriptor have been added:

    event->mode(int fd, int mode, int behavior);

The modes allowed are the same for the add() method but it adds:

    - DUDA_EVENT_WAKEUP

this one will only work if the file descriptor is in a sleep mode. The
behavior are the same described previously.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>