Response: new methods wait() and cont()
This patch implements two new methods in the response object to
handle different states of the connection.

 - response->wait(): instruct Duda to put in sleep mode the client
                     connection because the response will depend of
                     a third party callback.
 - response->cont(): continue a previous wait() connection, it
                     restore the client events.

Every time a wait() is set, a cont() must be invoked later in some
part of the web service.

On wait mode, no data is flushed until response->cont() and
response->end() are invoked.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>