Doc: New documentation generator
A new documentation prototype have been implemented. The documentation
style it's pretty much similar to Doxygen, due to the specific needs
of this project and the lack of time to learn Doxygen in deep, a new
script called 'generator' reads the source code files and generate the
documentation.

The documentation output is in HTML format and is powered by Twitter
Bootstrap.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>