Websocket: New channels support
this patch introduce the channels feature, now everytime a handshake
is performed for a new websocket incoming connection, this must be
associated to a specific channel number (or -1 to specify all). So
when the broadcast() and broadcast_all() methods are used, its
mandatory to specify to which channel the message will be broadcasted.

The new methods definition are:

   int handshake(duda_request_t *dr, int channel);

   int broadcast(ws_request_t *wr, unsigned char *data,
                 uint64_t len, int msg_type, int channel);

   int broadcast_all(unsigned char *data, uint64_t len,
                     int msg_type, int channel);

Signed-off-by: Eduardo Silva <edsiper@gmail.com>