Response: set TCP_CORK OFF after send the HTTP headers
After a callback, Duda send the HTTP headers through Monkey API,
but Monkey API per design let TCP_CORK ON in the socket in case
the caller wanted to send more data in the same round.

This patch disable TCP_CORK after send the headers and flush
some enqueued data.

Bonus track:
------------
The lesson of the day Feb 2, 2013: I must NEVER forget that when
sending the HTTP headers, Monkey sets the TCP_CORK flag ON in the
socket in case the caller wanted to send more data so we let know
the Kernel to buffer a little more bytes. If we do not set TCP_CORK
to OFF we will face some delays in the response. KeepAlive was very slow
due to this bug. More than 4 hours to found this silly bug.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>