Conf: fix string length variable for DocumentRoot key
When a DocumentRoot key is specified in the virtual host configuration
and it do not end with a slash, the slash is added at the end but it
was setting the wrong string length for the mk_pointer. Leading in a
problem when trying to map a static path inside the service.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>