Core: move API macros inside the structs
Before this patch, each API function that required to be expanded
through a macro by the pre-processor, was being declared as global,
this generated some issues when two API objects had similar function
names, in some cases the macro were overriden.

This patch put each API macro inside the struct, so we keep the references
protected to avoid any override.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>