Logger: new Log writer API
the Logger is a new API component that aims to provide communication channels
for all worker threads available, so they can write their own log messages. To
accomplish this from the stack point of view, a new worker thread monitor the
events on Kernel pipes and flush the logs every 3 seconds to disk.

The pending part at this point is to let the API use the LogDir configuration
directive to store the log messages.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>