Core: fixes when handling connections.
When a new request arrives to Duda I/O, it lookups for a current context
to try to re-use that slot allocation (mostly used on KeepAlive sessions).

There is a un-reproducible bug where sometimes the lookup returns
a duda_request_t that belongs to an old connection so it contains invalid
refereced data. I still cannot trap from where it comes from, as a workaround
for every service_run() call, we re-set the Monkey client_session (cs) and
session_request (sr) to the duda_request_t context.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>