Core: new 'service root' feature.
When a request arrives to Duda, it tries to match to which Virtual Host
should be assigned, after that it check which Web Services are registered
under that Virtual Host. So having multiple Web Services under the same
Virtual Host is possible defining different URI addresses for each one
based on their short name, e.g:

   app1 -> http://host1:port/app1/
   app2 -> hhtp://host1:port/app2/

The 'service root' feature allow to define a service it selfs as
'Service Root', which means that it will own the Virtual Host and others
Web Services will not be used (despites they are loaded). So when using
this feature the URI address becomes root and the short name of the
web service is not longer required, so if 'app1' defines it self as
'service root' it will be accessed through:

   http://host1:port/

The way to use this feature can be through the configuration or through
a hard call in duda_main():

1. Web Service Configuration:

   on the Virtual Host configuration, set the 'Root' key with value 'on'
   or 'off', e.g:

   [WEB_SERVICE]
       Enable on
       Root   on

   setting the key to 'on' means the service is root.

2. Setup through duda_main():

   When starting the service invoke the following method:

     conf->service_root();

Both options are valid, just note that using both ways at same time the
option #2 will override any setup made by the setup on #1.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>