Core: cleanup duda_request_t session when an Event trap the exception.
The issue is that if a remote connection is closed and the owner is a set
of Events callbacks, when closing, it was not instructing the Monkey core
to release the original HTTP references. The issue is visible when using
WebSockets or any other similar application.

This patch adds a new flag to trace if the HTTP session was cleared or not,
so on Event close it will cleanup if applies.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>