Params: rewrite get() method to use new Router mechanism.
This patch reimplement how the get() method works. Since the introduction
of the new Router mechanism and the support of dynamic variables on the URL,
the get() method allows to retrieve values based on key names, e.g:

  Router:  /test/:name/about/
  URL   :  /test/eduardo/about/

when using the API:

  char *name = param->get(dr, "name")

The 'name' variable references a new memory allocation with the 'eduardo'
string on it. The buffer is auto-freed once the callback ends.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>