Core/Plugins: migrate code to align with new Monkey v1.6 API
This patch apply the changes required to compile Duda with the new Monkey v1.6
stack. It touch almost all files of the source code.

Some code sections have been commented and added the 'FIXME' key, mostly on places
that we need to test functionality.

This is still work in process, current state is: it compiles.

Signed-off-by: Eduardo Silva <eduardo@monkey.io>